[#3772] Add configuration for CR-based summoning mode
Adds an alternate mode to summoning allowing for specifying CR
and type of summoned creatures rather than a specific document
link. This CR can be a single value or a formula so it can change
with level.

This does not include the actual summoning functionality, since
that will be dependent on the compendium browser.